% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.snn.R
\name{cal.snn}
\alias{cal.snn}
\title{Calculate shared nearest neighbors matrix from similarity matrix}
\usage{
cal.snn(Smat, knn = 50, bin = 5000)
}
\arguments{
\item{Smat}{the similarity matrix.}

\item{knn}{number of neighbors. We generally set it to total cells*0.01.}

\item{bin}{Matrix are divided by bin to avoid very long vectors, It does not affect the results.}
}
\value{
Return a shared nearest neighbor matrix.
}
\description{
Calculate shared nearest neighbors matrix from similarity matrix
}
\examples{
\dontrun{
snn<-cal.snn(Smat=Smat,knn=50)
}

}
