% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim.reduce.R
\name{dim.reduce}
\alias{dim.reduce}
\title{Perform Eigenvalue decompostion.}
\usage{
\method{dim}{reduce}(Smat, neigs = 50)
}
\arguments{
\item{Smat}{the similarity matrix.}

\item{neigs}{the number of Eigen vectors to calculate.}
}
\value{
Return a list with Eigenvalues and Eigen vectors.
}
\description{
Perform Eigenvalue decompostion.
}
\examples{
\dontrun{
eigs<-dim.reduce(Smat,neigs=30)
}

}
