% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigs.correct.R
\name{eigs.correct}
\alias{eigs.correct}
\title{correct Eigen vectors.}
\usage{
eigs.correct(
  eigs,
  knn_mat,
  batch,
  reference,
  knn_transfer_correction = 10,
  anchor_threshold = 5
)
}
\arguments{
\item{eigs}{the Eigen vectors.}

\item{knn_mat}{the knn matrix, output from eigs.knn.}

\item{batch}{the batch information.}

\item{reference}{datasets used as reference.}

\item{knn_transfer_correction}{the number of anchors used by non-anchor cells to learn their corrections.}

\item{anchor_threshold}{the minimum number of neighbors of an anchor cell.}
}
\value{
Return corrected Eigen vectors.
}
\description{
correct Eigen vectors.
}
\examples{
\dontrun{
eigs_corrected<-eigs.correct(eigs=eigs$vectors,
                            knn_mat=knn_update,
                            batch=batch,
                            reference="Resting",
                            knn_transfer_correction=10)
}

}
