% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigs.knn.R
\name{eigs.knn}
\alias{eigs.knn}
\title{learn neighbors across multiple datasets}
\usage{
eigs.knn(
  Smat,
  features,
  batch,
  reference,
  knn_target = 50,
  knn_reference = 20,
  threshold = 2
)
}
\arguments{
\item{Smat}{the similarity matrix.}

\item{features}{the guiding features.}

\item{batch}{the batch information.}

\item{reference}{which dataset in batch is used as reference. Reference should contain all cell types. If there are more than one references (A, B and C), we can specify reference=c("A","B","C") to align B to A, and align C to A and B.}

\item{knn_target}{number of neighbors for cells in target datasets (non-reference datasets) to learn in each reference.}

\item{knn_reference}{number of neighbors for cells in references to learn in each target dataset.}

\item{threshold}{the Z-score threshold used to filter false neighbors.}
}
\value{
Return a list containing knn matrix. It is the input of eigs.correct.
}
\description{
learn neighbors across multiple datasets
}
\examples{
\dontrun{
knn_update<-eigs.knn(Smat=Smat,
                    features=guide_features,
                    batch=batch,
                    reference="Resting",
                    knn_target=50,
                    knn_reference=10,
                    threshold=2)
}

}
