% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiConv_funs.R
\name{epiConv.anchor}
\alias{epiConv.anchor}
\title{Find nearest neighbors between batches}
\usage{
epiConv.anchor(
  Smat,
  row_sample,
  batch,
  reference,
  neigs = 30,
  features = list(),
  knn_target = 50,
  knn_reference = 20,
  threshold = 2
)
}
\arguments{
\item{Smat}{the similarity matrix.}

\item{row_sample}{indices of sampled cells for rows.}

\item{batch}{the batch information of cells.}

\item{reference}{the name of reference batch. Cells are sequentially aligned according the order of this vector.}

\item{neigs}{the number of Eigen vectors to calculate.}

\item{features}{features used to refine the knn matrix. It should be NULL if no feature is provided or a named list, where each element is a cell by feature matrix and the name is equal to the corresponding batch.}

\item{knn_target}{the number of nearest neighbors for cells from  query batch to find in reference batch.}

\item{knn_reference}{the number of nearest neighbors for cells from reference batch to find in query batch.}

\item{threshold}{the Z-score threshold to filter false neighbors.}
}
\value{
Return a list contains knn matrix between batches and associated data.
}
\description{
Find nearest neighbors between batches
}
\examples{
\dontrun{
expr<-matrix(runif(10000),100,100)
PCs<-prcomp(expr)$x[,1:20]
res_anchor<-epiConv.anchor(Smat=Smat,
                           row_sample=1:ncol(Smat),
                           batch=batch,
                           reference=c("Batch1","Batch2"),
                           neigs=30,
                           features=list(Batch1=PCs),
                           knn_target=50,
                           knn_reference=20,
                           threshold=2)

}

}
