% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiConv_funs.R
\name{epiConv.correct.parallel}
\alias{epiConv.correct.parallel}
\title{Correct batch effects on similarity matrix in parallel}
\usage{
epiConv.correct.parallel(
  Smat,
  row_sample,
  batch,
  reference,
  neigs = 30,
  knn_update,
  ncore = 1
)
}
\arguments{
\item{Smat}{the similarity matrix.}

\item{row_sample}{indices of sampled cells for rows.}

\item{batch}{the batch information of cells.}

\item{reference}{the name of reference batch. Cells are sequentially aligned according the order of this vector.}

\item{neigs}{the number of Eigen vectors to calculate.}

\item{knn_update}{the knn matrix between batches.}

\item{ncore}{number of threads.}
}
\value{
Return a list contains Eigen vectors before and after correction.
}
\description{
Correct batch effects on similarity matrix in parallel
}
\examples{
\dontrun{
res_correct<-epiConv.correct.parallel(Smat=Smat,
                                      row_sample=1:ncol(Smat),
                                      batch=batch,
                                      reference=c("Batch1","Batch2"),
                                      neigs=30,
                                      knn_update=res_anchor$knn_update,
                                      ncore=5)

}

}
