% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiConv.joint.R
\name{epiConv.joint}
\alias{epiConv.joint}
\title{Find nearest neighbors for each column}
\usage{
epiConv.joint(
  Smat,
  batch,
  reference,
  neigs = 30,
  features = list(),
  knn_mat = list(),
  knn_target = 50,
  knn_reference = 10,
  threshold = 2
)
}
\arguments{
\item{Smat}{the similarity matrix.}

\item{batch}{the batch information of cells.}

\item{reference}{the name of reference batch. Cells are sequentially aligned according the order of this vector.}

\item{neigs}{the number of Eigen vectors to calculate.}

\item{features}{features used to refine the knn matrix. It should be NULL if no feature is provided or a named list contains the features of cells from some batches.}

\item{knn_mat}{the pre-computed knn_mat. It should be NULL if no knn_mat is provided or a named list contains the knn_mat of some batches.}

\item{knn_target}{the number of nearest neighbors for cells from  query batch to find in reference batch.}

\item{knn_reference}{the number of nearest neighbors for cells from reference batch to find in query batch.}

\item{threshold}{the Z-score threshold to filter false neighbors.}
}
\value{
Return a list contains Eigen vectors before correction, after correction and knn matrix across batches.
}
\description{
Find nearest neighbors for each column
}
\examples{
\dontrun{
res_joint<-epiConv.joint(Smat=Smat,
                         batch=batch,
                         reference=c("ref1","ref2"),
                         neigs=30,
                         features=NULL,
                         knn_mat=NULL,
                         knn_target=50,
                         knn_reference=10,
                         threshold=2)

}

}
