% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiConv_funs.R
\name{epiConv.louvain}
\alias{epiConv.louvain}
\title{Perform louvain clustering on SNN matrix}
\usage{
epiConv.louvain(snn, resolution, ...)
}
\arguments{
\item{snn}{the SNN matrix.}

\item{resolution}{the resoltions for louvain clustering.}
}
\value{
clustering results
}
\description{
Perform louvain clustering on SNN matrix
}
\examples{
\dontrun{
clust<-epiConv.louvain(snn=snn_mat,
                       resolution=c(0.8,0.6,0.4,0.2))
head(clust)

}

}
