% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.knn.R
\name{find.knn}
\alias{find.knn}
\title{Find nearest neighbors for each column}
\usage{
find.knn(Smat, knn, binarize = T)
}
\arguments{
\item{Smat}{the similarity matrix.}

\item{knn}{number of nearest neighbors.}

\item{binarize}{if binarize=T, the function will return 1 for a neighbor or 0 otherwise. if binarize=F, the function will return the similarity value for a neighbor or 0 otherwise}
}
\value{
Return a knn matrix.
}
\description{
Find nearest neighbors for each column
}
\examples{
\dontrun{
knn<-find.knn(Smat=Smat,
              knn=20,
              binarize=T)
}

}
