% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.estimate.R
\name{freq.estimate}
\alias{freq.estimate}
\title{Calculate frequency of peaks}
\usage{
freq.estimate(mat)
}
\arguments{
\item{mat}{matrix.}
}
\value{
Return a vector with frequency of peaks.
}
\description{
Calculate frequency of peaks
}
\examples{
\dontrun{
freq<-freq.estimate(mat)
}

}
