% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.estimate.R
\name{lib.estimate}
\alias{lib.estimate}
\title{Calculate number of accessible peaks of cells}
\usage{
lib.estimate(mat)
}
\arguments{
\item{mat}{matrix.}
}
\value{
Return a vector with number of accessible of cells
}
\description{
Calculate number of accessible peaks of cells
}
\examples{
\dontrun{
lib_size<-lib.estimate(mat)
}

}
