% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.epiConv.R
\name{run.epiConv}
\alias{run.epiConv}
\title{Calculate similarities between cells}
\usage{
run.epiConv(
  mat,
  lib_size,
  nbootstrap = 15,
  nsample = floor(nrow(mat) * 0.2),
  bin = 10000,
  inf_replace = (-8)
)
}
\arguments{
\item{mat}{Matrix object constains the peak by cell matrix.}

\item{lib_size}{library size of single cells.}

\item{nbootstrap}{number of bootstraps performed.}

\item{nsample}{number of peaks sampled in each bootstrap.}

\item{bin}{Matrix are divided by bin to avoid very long vectors, It does not affect the results.}

\item{inf_replace}{a small value used to replace -Inf.}
}
\value{
Return a similarity matrix.
}
\description{
Calculate similarities between cells
}
\examples{
\dontrun{
Smat<-run.epiConv(mat=mat,
                  lib_size=colSums(mat),
                  nbootstrap=15,
                  nsample=floor(nrow(mat)*0.2),
                  bin=5000,
                  inf_replace=(-8))
}

}
