% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiConv_funs.R
\name{Smat2snn}
\alias{Smat2snn}
\title{calculate shared-nearest-neighbor(SNN) matrix}
\usage{
Smat2snn(Smat, knn, ncore = 1)
}
\arguments{
\item{Smat}{the similarity matrix.}

\item{knn}{number of nearest neighbors kept.}

\item{ncore}{number of threads.}
}
\value{
a sparse matrix
}
\description{
calculate shared-nearest-neighbor(SNN) matrix
}
\examples{
\dontrun{
snn_mat<-Smat2snn(Smat=Smat,
                  knn=20,
                  ncore=5)

}

}
