% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiConv_funs.R
\name{lib.estimate}
\alias{lib.estimate}
\title{Calculate number of accessible peaks of cells}
\usage{
lib.estimate(mat)
}
\arguments{
\item{mat}{a dgCMatrix object.}
}
\value{
Return a vector of library sizes
}
\description{
Calculate number of accessible peaks of cells
}
\examples{
\dontrun{
lib_size<-lib.estimate(mat)
}

}
